native: JNI runtimeBrowseRespondToHostKey の bool→u8 変換漏れを修正 (#27)
jboolean(jni 0.22 では bool) を u8 引数の ffi にそのまま渡していて
Android ターゲットのみで E0308 コンパイル失敗。他の jboolean 境界と
同じく u8::from(accept) に修正。ホスト側 cfg ではコンパイルされず
Android ビルドでのみ顕在化していた。

+++ native/meeterm-core/src/jni.rs
@@ -1004,7 +1004,7 @@ pub extern "system" fn Java_dev_meeterm_terminal_MeetermNative_runtimeBrowseResp
                 token.len(),
                 fingerprint.as_ptr(),
                 fingerprint.len(),
-                accept,
+                u8::from(accept),
             )
         })
     }))
